Auth

+++ src/containers/AdminMural.js
@@ -1,9 +1,10 @@
 // dependencies
-import React, { useEffect, useState } from "react";
+import React, { useContext, useEffect, useState } from "react";
 import styled from "styled-components";
 import axios from "axios";
 // components
 import { AdminMuralForm, AdminPageTitle, LoadingSpinner } from "../components";
+import AdminHeader from "./AdminHeader";
 import {
     StyledGreenSoloButton,
     StyledOutlineButton,
@@ -12,6 +13,9 @@ import { StyledP } from "./styles/adminTypography";
 // hooks
 import { useForm } from "react-hook-form";
 import { useHistory, useParams } from "react-router";
+import { useAuth0 } from "@auth0/auth0-react";
+// context
+import MenuContext from "../context/MenuContext";
 
 const StyledAdminMural = styled.div`
     display: flex;
@@ -88,6 +92,10 @@ const AdminMural = ({ variant }) => {
     } = useForm();
     // set up history
     const history = useHistory();
+    // set up hooks
+    const { isLoading, getAccessTokenSilently } = useAuth0();
+    // set up context
+    const { menuOpen } = useContext(MenuContext);
 
     useEffect(() => {
         async function getMuralData() {
@@ -117,12 +125,19 @@ const AdminMural = ({ variant }) => {
         // on form submit, send data to API, move to next step
         setLoading(true);
         try {
+            const token = await getAccessTokenSilently();
+
             if (variant === "Add") {
                 const res = await axios.post(
                     process.env.REACT_APP_BACKEND_URL + "murals/",
                     {
                         title: data.title,
                         description: data.description,
+                    },
+                    {
+                        headers: {
+                            Authorization: `Bearer ${token}`,
+                        },
                     }
                 );
                 const id = res.data.mural.id;
@@ -131,7 +146,12 @@ const AdminMural = ({ variant }) => {
                 await axios.patch(
                     process.env.REACT_APP_BACKEND_URL +
                         `murals/mural/${muralId}`,
-                    data
+                    data,
+                    {
+                        headers: {
+                            Authorization: `Bearer ${token}`,
+                        },
+                    }
                 );
             }
             setError(null);
@@ -153,10 +173,17 @@ const AdminMural = ({ variant }) => {
         let res;
 
         try {
+            const token = await getAccessTokenSilently();
+
             res = await axios.post(
                 process.env.REACT_APP_BACKEND_URL +
                     `murals/upload/${muralId}/image/${step}`,
-                formData
+                formData,
+                {
+                    headers: {
+                        Authorization: `Bearer ${token}`,
+                    },
+                }
             );
             if (step < 3) {
                 reset("", { keepValues: false });
@@ -188,9 +215,19 @@ const AdminMural = ({ variant }) => {
         // push admin to Admin Murals page
         setLoading(true);
         try {
-            await axios.delete(
-                process.env.REACT_APP_BACKEND_URL + `murals/mural/${muralId}`
-            );
+            const token = await getAccessTokenSilently();
+
+            if (step > 0) {
+                await axios.delete(
+                    process.env.REACT_APP_BACKEND_URL +
+                        `murals/mural/${muralId}`,
+                    {
+                        headers: {
+                            Authorization: `Bearer ${token}`,
+                        },
+                    }
+                );
+            }
             history.push("/admin/murals");
         } catch (err) {
             console.log("Cancel Error", err);
@@ -206,67 +243,80 @@ const AdminMural = ({ variant }) => {
             </StyledAdminMural>
         );
 
-    return (
-        <StyledAdminMural>
-            <AdminPageTitle>{variant} Mural</AdminPageTitle>
-            {step > 0 && <StyledText>Upload Photo #{step} of 3</StyledText>}
-            {error && <StyledError>{error}</StyledError>}
-            {step === 0 && variant === "Add" && (
-                <AdminMuralForm
-                    handleDataSubmit={handleDataSubmit}
-                    handleCancel={handleCancel}
-                    variant={variant}
-                />
-            )}
-            {step === 0 && variant === "Edit" && (
-                <AdminMuralForm
-                    handleDataSubmit={handleDataSubmit}
-                    handleCancel={handleCancel}
-                    variant={variant}
-                    preloadedValues={preloadedValues}
-                />
-            )}
-            {step > 0 && (
-                <>
-                    <StyledUploadForm
-                        onSubmit={handleSubmit(handleImageSubmit)}
-                    >
-                        <StyledUploadInput
-                            type="file"
-                            id="upload"
-                            name="upload"
-                            {...register("upload", {
-                                required: "Image is required",
-                            })}
+    return menuOpen ? (
+        <AdminHeader />
+    ) : (
+        <>
+            {" "}
+            <AdminHeader />{" "}
+            <main>
+                <StyledAdminMural>
+                    <AdminPageTitle>{variant} Mural</AdminPageTitle>
+                    {step > 0 && (
+                        <StyledText>Upload Photo #{step} of 3</StyledText>
+                    )}
+                    {error && <StyledError>{error}</StyledError>}
+                    {step === 0 && variant === "Add" && (
+                        <AdminMuralForm
+                            handleDataSubmit={handleDataSubmit}
+                            handleCancel={handleCancel}
+                            variant={variant}
                         />
-                        <StyledSubmitBtn>Submit</StyledSubmitBtn>
-                        {variant === "Edit" && (
-                            <StyledContinueBtn
-                                onClick={handleContinue}
-                                color="#207070"
+                    )}
+                    {step === 0 && variant === "Edit" && (
+                        <AdminMuralForm
+                            handleDataSubmit={handleDataSubmit}
+                            handleCancel={handleCancel}
+                            variant={variant}
+                            preloadedValues={preloadedValues}
+                        />
+                    )}
+                    {step > 0 && (
+                        <>
+                            <StyledUploadForm
+                                onSubmit={handleSubmit(handleImageSubmit)}
                             >
-                                Keep Image
-                            </StyledContinueBtn>
-                        )}
-                        <StyledCancelBtn color="#DB9487" onClick={handleCancel}>
-                            Cancel (Deletes Mural)
-                        </StyledCancelBtn>
-                    </StyledUploadForm>
-                </>
-            )}
-            {step > 0 && variant === "Edit" && (
-                <>
-                    <StyledText>Current Image:</StyledText>
-                    <StyledImg
-                        src={
-                            process.env.REACT_APP_BACKEND_URL +
-                            `murals/mural/${muralId}/image/${step}`
-                        }
-                        alt="Mural"
-                    />
-                </>
-            )}
-        </StyledAdminMural>
+                                <StyledUploadInput
+                                    type="file"
+                                    id="upload"
+                                    name="upload"
+                                    {...register("upload", {
+                                        required: "Image is required",
+                                    })}
+                                />
+                                <StyledSubmitBtn>Submit</StyledSubmitBtn>
+                                {variant === "Edit" && (
+                                    <StyledContinueBtn
+                                        onClick={handleContinue}
+                                        color="#207070"
+                                    >
+                                        Keep Image
+                                    </StyledContinueBtn>
+                                )}
+                                <StyledCancelBtn
+                                    color="#DB9487"
+                                    onClick={handleCancel}
+                                >
+                                    Cancel (Deletes Mural)
+                                </StyledCancelBtn>
+                            </StyledUploadForm>
+                        </>
+                    )}
+                    {step > 0 && variant === "Edit" && (
+                        <>
+                            <StyledText>Current Image:</StyledText>
+                            <StyledImg
+                                src={
+                                    process.env.REACT_APP_BACKEND_URL +
+                                    `murals/mural/${muralId}/image/${step}`
+                                }
+                                alt="Mural"
+                            />
+                        </>
+                    )}
+                </StyledAdminMural>
+            </main>
+        </>
     );
 };
